scripts/requirements.txt
pandas==2.2.3
python-dotenv==1.1.0
supabase==2.10.0
pypdf==5.1.0
requests==2.32.3
